feat: Add Node class
This class holds node data and has some functions that modify,
serialize, and deserialize node data.

+++ src/main/java/org/tasker/Rect.java
@@ -1,13 +1,13 @@
 package org.tasker;
 
 class Rect {
-  public int x, y, w, h;
+  public double x, y, w, h;
 
   public boolean contains(Vec2 p) {
     return p.x >= x && p.x <= x + w && p.y >= y && p.y <= y + h;
   }
 
-  public Rect(int x, int y, int w, int h) {
+  public Rect(double x, double y, double w, double h) {
     this.x = x;
     this.y = y;
     this.w = w;
