+++ src/main/java/org/tasker/Vec2.java
@@ -1,9 +1,9 @@
 package org.tasker;
 
 class Vec2 {
-  public int x, y;
+  public double x, y;
 
-  public Vec2(int x, int y) {
+  public Vec2(double x, double y) {
     this.x = x;
     this.y = y;
   }
